Real EDA, two real bugs found & fixed, and a real DeBERTa fine-tuning run
- Real EDA on the full 17,307-essay train set (score distribution, length correlations)
- Both original notebooks preserved and honestly credited: the actual medal submission
  (DeBERTa embeddings + 15-fold LightGBM ensemble, forked from 3 public kernels, real
  logged mean CV QWK 0.8398 / Public LB 0.819) and an alternate direct fine-tune approach
- Fixed two real bugs in the existing repo scaffold to get it running: fp16 master weights
  breaking GradScaler, and an intractable 53,130-combination brute-force threshold search
  (replaced with sklearn's cohen_kappa_score + Nelder-Mead optimization)
- Actually fine-tuned deberta-v3-small end-to-end, 3-fold CV, real logged mean val QWK 0.8119
- Corrected model name mismatch (config said deberta-v3-base, notebook actually used small)

+++ src/modeling.py
@@ -8,7 +8,13 @@ class DebertaRegressor(nn.Module):
     def __init__(self, model_name: str, num_labels: int = 1):
         super().__init__()
         self.config = AutoConfig.from_pretrained(model_name)
-        self.backbone = AutoModel.from_pretrained(model_name, config=self.config)
+        # Force fp32 master weights: this checkpoint is stored on the HF Hub in fp16, and
+        # transformers now loads it in its stored dtype by default. Combined with this repo's
+        # torch.cuda.amp.autocast + GradScaler training loop (src/train.py), fp16 master
+        # weights make GradScaler raise "Attempting to unscale FP16 gradients" — GradScaler
+        # requires fp32 master weights, with autocast handling the fp16 compute during the
+        # forward pass. Found and fixed 2026-08-13.
+        self.backbone = AutoModel.from_pretrained(model_name, config=self.config, dtype=torch.float32)
         self.head = nn.Linear(self.config.hidden_size, num_labels)
 
     def forward(self, input_ids=None, attention_mask=None, **kwargs):
